requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-Migrate==4.0.4
WTForms==3.0.1
email-validator==2.0.0
python-dotenv==1.0.0
psycopg2-binary==2.9.6
instagrapi==2.0.0
requests==2.31.0
Pillow==9.5.0
gunicorn==21.2.0
SQLAlchemy==2.0.19
alembic==1.11.1
Werkzeug==2.3.6
cryptography==41.0.0
openai==1.6.1
feedparser==6.0.10
moviepy==1.0.3
imageio-ffmpeg==0.4.9
